pyproject.toml
[tool.poetry]
name = "rental-property-estimator"
version = "0.1.0"
description = ""
authors = ["zainaraza43 <urboyzain@gmail.com>"]
readme = "README.md"
packages = [{include = "rental_property_estimator"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
ratelimit = "^2.2.1"
python-dotenv = "^1.0.0"
datetime = "^5.2"
sqlalchemy = "^2.0.20"
py-cord = "^2.4.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.287"
black = "^23.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
